+++ libsolidity/interface/CompilerStack.cpp
@@ -28,6 +28,7 @@
 
 #include <libsolidity/analysis/ControlFlowAnalyzer.h>
 #include <libsolidity/analysis/ControlFlowGraph.h>
+#include <libsolidity/analysis/ControlFlowRevertPruner.h>
 #include <libsolidity/analysis/ContractLevelChecker.h>
 #include <libsolidity/analysis/DeclarationTypeChecker.h>
 #include <libsolidity/analysis/DocStringAnalyser.h>
@@ -516,10 +517,12 @@ bool CompilerStack::analyze()
 
 			if (noErrors)
 			{
+				ControlFlowRevertPruner pruner(cfg);
+				pruner.run();
+
 				ControlFlowAnalyzer controlFlowAnalyzer(cfg, m_errorReporter);
-				for (Source const* source: m_sourceOrder)
-					if (source->ast && !controlFlowAnalyzer.analyze(*source->ast))
-						noErrors = false;
+				if (!controlFlowAnalyzer.run())
+					noErrors = false;
 			}
 		}